{"schema":"decodex/commit/1","summary":"Split consolidation proposal DTOs into a child module after strict validation.","authority":"manual"}
